gui fix

+++ osinstall/tumbleweed/modules/gui.sh
@@ -65,8 +65,11 @@ flatpak install -y flathub com.github.tchx84.Flatseal
 bash $SCRIPT_DIR/../../derivations/yt.sh
 
 # chrome
-mkdir -p /opt/rpms/
-rpm --import https://dl.google.com/linux/linux_signing_key.pub
-wget -O /opt/rpms/google-chrome-stable_current_x86_64.rpm https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm
+if [ ! -f /opt/rpms/google-chrome-stable_current_x86_64.rpm ];then
+    mkdir -p /opt/rpms/
+    rpm --import https://dl.google.com/linux/linux_signing_key.pub
+    wget -O /opt/rpms/google-chrome-stable_current_x86_64.rpm https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm
+fi
+
 zypper install -y /opt/rpms/google-chrome-stable_current_x86_64.rpm
